mendora-ai/requirements.txt
streamlit
google-generativeai
pandas
plotly
pypdf
langchain
chromadb
sentence-transformers
